Move wyre widget code to vendor directory
