+++ ui/app/components/app/modals/deposit-ether-modal.js
@@ -5,7 +5,7 @@ const inherits = require('util').inherits
 const connect = require('react-redux').connect
 const actions = require('../../../store/actions')
 const { getNetworkDisplayName } = require('../../../../../app/scripts/controllers/network/util')
-const openWyre = require('../../../../lib/wyre')
+const openWyre = require('../../../../vendor/wyre')
 const { DEPOSIT_ROUTE } = require('../../../helpers/constants/routes')
 const { ENVIRONMENT_TYPE_POPUP } = require('../../../../../app/scripts/lib/enums')
 const { getEnvironmentType } = require('../../../../../app/scripts/lib/util')
